+++ upcloud/resource_upcloud_price_zone.go
@@ -1,7 +1,7 @@
 package upcloud
 
 import (
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
 func resourceUpCloudPriceZone() *schema.Resource {